+++ app/page.js
@@ -1,5 +1,10 @@
 import CRMManager from "@/components/CRMManager";
+import AuthGate from "@/components/AuthGate";
 
 export default function Home() {
-  return <CRMManager />;
+  return (
+    <AuthGate>
+      <CRMManager />
+    </AuthGate>
+  );
 }